requirements.txt
python-telegram-bot==13.7
urllib3==1.26.15
pytz
certifi
six
pillow==10.0.0
pywin32==306; platform_system=="Windows"
plyer==2.1.0
customtkinter==5.2.0 